requirements.txt
Flask>=2.0.0
supabase>=2.0.0
python-dotenv>=0.19.0 # Add dotenv for .env file loading
requests>=2.25.0 # For making HTTP requests to APIs

# Add dotenv for .env file loading
